Update floats to ints.

+++ module10/HelloWorld.cl
@@ -28,32 +28,32 @@ __kernel void mult_kernal(__global const float *a,
     result[gid] = a[gid] * b[gid];
 }
 
-// A kernal for "oring" the values in a and b into result.
+// A kernal for "oring" the values integer values of a and b into result.
 __kernel void or_kernal(__global const float *a,
 						__global const float *b,
 						__global float *result)
 {
     int gid = get_global_id(0);
 
-    result[gid] = a[gid] | b[gid];
+    result[gid] = static_cast<int>(a[gid]) | static_cast<int>(b[gid]);
 }
 
-// A kernal for "anding" the values in a and b into result.
+// A kernal for "anding" the values integer values of a and b into result.
 __kernel void and_kernal(__global const float *a,
 						__global const float *b,
 						__global float *result)
 {
     int gid = get_global_id(0);
 
-    result[gid] = a[gid] & b[gid];
+    result[gid] = static_cast<int>(a[gid]) & static_cast<int>(b[gid]);
 }
 
-// A kernal for "exclusing oring" the values in a and b into result.
+// A kernal for "exclusing oring" the values integer values of a and b into result.
 __kernel void xor_kernal(__global const float *a,
 						__global const float *b,
 						__global float *result)
 {
     int gid = get_global_id(0);
 
-    result[gid] = a[gid] ^ b[gid];
+    result[gid] = static_cast<int>(a[gid]) ^ static_cast<int>(b[gid]);
 }
